structure updated as in readme
